docs: rename extension
