+++ .docs/extensions/automatic_redirects.py
